Attempting to make bomb path feature
At a stage with the bomb path feature where it works as expected if the
finish node is further away than the bomb node and if the path finding
algorithm doesn't have to "backtrack", which is not the exact
intentions. TODO: fix path finding algorithms for when a bomb is on the
grid.