Reverted CSS performance optimization changes
Reverted the CSS performance optimization changes because it messed with
the original visuals of the wanted animations and the performance gains
were very minimal. The "heavier" performance simply as has to do with
the fact that the project is built on Next.JS with React. It will never
be completely lightweight, unfortunately.