Removed unused util function
Removed the getNewGridWithWeightToggled function as it was no longer
being used. We exclusively create weights from the basic weights maze
and no other way. This is due to performance inconsistencies with
manually drawing weights like we do with walls. For some reason it's
just incredibly janky, laggy, inconsistent and terrible user experience.