Added feature to draw weight nodes on the grid
In addition to be able to create a weight node maze for the grid, you
can now also draw and place your own weight nodes as you see fit on the
grid. Weight nodes have also changed to have static weights for
consistency and better comparison analysis. They now have a weight of
either 10 or 15, depending on your screen's viewport. Before they had a
weight of 2-10 or 2-15 depending on viewport. This randomization has
been removed.